c++: add -fdump-lang-tinst
This patch adds a dump with a trace of template instantiations, indented
based on the depth of recursive instantiation. -lineno adds the location
that triggered the instantiation, -details adds non-instantiation
sbustitutions.

The instantiate_pending_templates change is to avoid a bunch of entries for
reopening tinst scopes that we then don't instantiate anything with; it also
seems a bit cleaner this way.

gcc/cp/ChangeLog:

	* cp-tree.h: Declare tinst_dump_id.
	* cp-objcp-common.cc (cp_register_dumps): Set it.
	* pt.cc (push_tinst_level_loc): Dump it.
	(reopen_tinst_level): Here too.
	(tinst_complete_p): New.
	(instantiate_pending_templates): Don't reopen_tinst_level for
	already-complete instantiations.

gcc/ChangeLog:

	* doc/invoke.texi: Move C++ -fdump-lang to C++ section.
	Add -fdump-lang-tinst.