testsuite: arm: remove arm32 check from a few effective-targets
A few arm effective-targets call check_effective_target_arm32 even
though they would force a -march=XXX flag which supports Arm and/or
Thumb-2, thus making the arm32 check useless.  This has an impact when
the toolchain is configured with a default -march or -mcpu which
supports Thumb-1 only: in such a case, arm32 is false and we skip many
tests, thus reducing coverage.

This patch removes the call to check_effective_target_arm32 where it
is useless, enabling about 2000 tests.

In addition, add an early exit if the target is not an arm one, thus
saving a few compilation cycles where not needed.  In all callers of
arm_neon_ok, remove the now useless "istarget arm*-*-*.

gcc/testsuite/ChangeLog:

	* lib/target-supports.exp
	(check_effective_target_arm_neon_ok_nocache): Remove arm32 check.
	Add istarget arm*-*-* check.
	(check_effective_target_arm_neon_fp16_ok_nocache): Likewise.
	(check_effective_target_arm_neon_softfp_fp16_ok_nocache): Likewise.
	(check_effective_target_arm_v8_neon_ok_nocache): Likewise.
	(check_effective_target_arm_neonv2_ok_nocache): Likewise.
	(check_effective_target_vect_pack_trunc): Remove istarget arm*-*-*
	check.
	(check_effective_target_vect_unpack): Likewise.
	(check_effective_target_vect_condition): Likewise.
	(check_effective_target_vect_cond_mixed): Likewise.
	(available_vector_sizes): Likewise.