Fix pr118947-1.c and pr78408-3.c on targets where 32 bytes memcpy uses a vector
The problem here is on targets where a 32byte memcpy will use an integral (vector) type
to do the copy and the code will be optimized a different way than expected. This changes
the testcase instead to use a size of 1025 to make sure there is no target that will use an
integral (vector) type for the memcpy and be optimized via the method that was just added.

Pushed as obvious after a test run.

gcc/testsuite/ChangeLog:

	* gcc.dg/pr118947-1.c: Use 1025 as the size of the buf.
	* gcc.dg/pr78408-3.c: Likewise.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>