[RISC-V][PR target/119865] Don't free ggc allocated memory
Kaiweng's patch to stop freeing riscv_arch_string was correct, but incomplete
as there's another path that was freeing that node, which is just plain wrong
for a node allocated by the GC system.

This patch removes that call to free() which fixes the test.  I've spun it in
my tester and will obviously wait for the pre-commit system to render a verdict
before moving forward.

	PR target/119865
gcc/
	* config/riscv/riscv.cc (parse_features_for_version): Do not
	explicitly free the architecture string.