forwprop: Fix looping after fold_stmt and some forwprop local folds happen
r10-2587-gcc19f80ceb27cc added a loop over the current statment if there was
a change. Except in some cases it turns out changed will turn from true to false
because instead of doing |= after the fold_stmt, there was an just an `=`.
This fixes that and now we loop even if fold_stmt changed the statement and
there was a local fold that happened.

gcc/ChangeLog:

	* tree-ssa-forwprop.cc (pass_forwprop::execute): Use `|=` for
	changed on the local folding.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>