gimple-fold: Don't replace `{true/false} != false` with `true/false` inside GIMPLE_COND
This is like the patch where we don't want to replace `bool_name != 0`
with `bool_name` but for instead for INTEGER_CST. The only thing
difference is there are a few different forms for always true/always
false; only handle it if it was in the canonical form. A few new helpers are
added for the canonical form detection.

This also replaces the previous version of the patch which did an early
exit from fold_stmt_1 instead so we can change the non-canonical form
into a canonical in the end.

gcc/ChangeLog:

	* gimple.h (gimple_cond_true_canonical_p): New function.
	(gimple_cond_false_canonical_p): New function.
	* gimple-fold.cc (replace_stmt_with_simplification): Return
	false if replacing the operands of GIMPLE_COND with an INTEGER_CST
	and already in canonical form.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>