ada: Filling in gaps in handling of inherited Pre/Post'Class aspects
The initial set of changes for doing proper mapping of calls to primitive
functions in Pre/Post'Class aspects inherited by derived types was not
handling some cases (such as when formals are referenced as part of
dereferences, certain aspects such as 'Old and 'Access, and conditional
and declare expressions), and mishandling other cases (such as nested
function calls).

This set of changes attempts to properly address those cases. It also
includes a change to suppress unneeded (and sometimes wrong) accessibility
checks on conversions of actual parameters of a derived type to the parent
type when passing them on calls to parent primitives (encountered while
developing these changes).

gcc/ada/ChangeLog:

	* exp_util.adb (Must_Map_Call_To_Parent_Primitive): Change function
	name (was Call_To_Parent_Dispatching_Op_Must_Be_Mapped). Move logic
	for attributes and dereferences, plus testing for controlled formals,
	into new function Expr_Has_Ctrl_Formal_Ref. Add handling for
	access attributes, multiple levels of attributes/dereferences,
	conditional_expressions, and declare_expressions. Properly account
	for function calls with multiple operands and enclosing calls.
	(Expr_Has_Ctrl_Formal_Ref): New function to determine whether
	an expression is a reference to a controlling formal or has
	a prefix that is such a reference.
	(Is_Controlling_Formal_Ref): New function in Expr_Has_Ctrl_Formal_Ref
	to determine if a node is a direct reference to a controlling formal.
	* freeze.adb (Build_DTW_Body): Create an unchecked conversion instead
	of a regular type conversion for converting actuals in calls to parent
	inherited primitives that are wrapped for inherited pre/postconditions.
	Avoids generating unnecessary checks (such as accessibility checks on
	conversions for anonymous access formals).