gimple-verifier: Add check that comparison in GIMPLE_COND does not throw
While working on PR 119903, I noticed that there is code in replace_stmt_with_simplification
which makes sure that the comparison of a GIMPLE_COND does not throw
(non-call exceptions and trapping math) but the gimple verifier does not
verify this. So let's add it.

Bootstrapped and tested on x86_64-linux-gnu.

gcc/ChangeLog:

	* tree-cfg.cc (verify_gimple_cond): Error out if the comparison
	throws.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>