c++/modules: Ensure DECL_FRIEND_CONTEXT is streamed [PR119939]
An instantiated friend function relies on DECL_FRIEND_CONTEXT being set
to be able to recover the template arguments of the class that
instantiated it, despite not being a template itself.  This patch
ensures that this data is streamed even when DECL_CLASS_SCOPE_P is not
true.

	PR c++/119939

gcc/cp/ChangeLog:

	* module.cc (trees_out::lang_decl_vals): Also stream
	lang->u.fn.context when DECL_UNIQUE_FRIEND_P.
	(trees_in::lang_decl_vals): Likewise.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/concept-11_a.H: New test.
	* g++.dg/modules/concept-11_b.C: New test.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>